pyproject.toml
[project]
name = "sesame-tts"
version = "1.0.0"
description = "OpenAI-compatible local TTS server wrapping Sesame CSM-1B via mlx-audio"
requires-python = ">=3.10,<3.14"
dependencies = [
    "mlx-audio>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[tool.uv]
prerelease = "allow"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "anyio[trio]>=4.0.0",
    "httpx>=0.27.0",
    "asgi-lifespan>=2.0.0",
]
voice-loop = [
    "mlx-whisper>=0.4.0",
    "sounddevice>=0.5.0",
    "webrtcvad>=2.0.10",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "rumps>=0.4.0",
    "PyYAML>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "setuptools>=82.0.0",
]
